requirements.txt
PyPDF2==1.26.0
rpaframework==11.5.1
selenium==3.141.0